Use EUI-48 format to show MAC addresses
Across `capwap-ac` we use a set of different, allthough correct,
variants to format the MAC address of entities:

* ieee80211_station:format_mac/1 uses lowercase hex-values, separated by
  colons.
* flower_tools:format_mac/1 uses uppercase hex-values, separated by
  colons.

An encoding to use for MAC addresses is EUI-48, as described in
http://standards.ieee.org/develop/regauth/tut/eui.pdf. The octets are
given uppercase and are separated via hyphens.

This commit uses the EUI-48 format to render MAC addresses in various
places, either for log messages or as part of AAA-sessions.