Properly handle spawned children inside a container
A process spawning children is usually responsible for taking care of
them if they stop working. Usually, that responsibility is handed over
to the parent process and thus ends with Pid 1.

When Erlang is running as Pid 1, it is not taking care of stopped child
processes and thus, over time, the process list fills up. Especially, if
an operator enters the remote console.

This commit adds "tini" as new Pid 1 to the container. "tini" will take
care about the stopped child processes.